beginner/submissions/team-members/sheila-adams/requirements.txt
# Contains all libraries that are needed for this project

# Data manipulation and analysis
# pandas
# numpy

# Visualization
# matplotlib
# seaborn
# plotly

# Machine learning
# scikit-learn
# xgboost
# lightgbm
# catboost

# Deep learning
# tensorflow
# keras
# torch
# torchvision

# Data processing
# scipy

# Jupyter support
jupyter
notebook
ipykernel

# Model evaluation
# statsmodels

# Utilities
# tqdm

# Optional: Data connectors
# sqlalchemy
# openpyxl

openai-agents
python-dotenv
tavily-python